separate checkbox and radio boolean pickers
