--- frontend/src/metabase/query_builder/components/filters/pickers/BooleanPicker/index.js
@@ -1,1 +1,0 @@
-export { default } from "./BooleanPicker";
