+++ frontend/src/metabase/query_builder/components/filters/pickers/BooleanPicker/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./BooleanPickerRadio";
+export { BooleanPickerCheckbox } from "./BooleanPickerCheckbox";
