+++ frontend/test/__support__/e2e/helpers/e2e-boolean-helpers.js
@@ -1,13 +1,10 @@
 // until we have a test dataset that includes boolean data, we can use this questions to test booleans
-const BOOLEAN_QUERY =
-  'select 0::integer as "integer", true::boolean AS "boolean" union all \nselect 1::integer as "integer", false::boolean AS "boolean" union all \nselect null as "integer", true::boolean AS "boolean" union all \nselect -1::integer as "integer", null AS "boolean"';
-
-export const setupBooleanQuery = () => {
+export function setupBooleanQuery(questionName = "Boolean Query") {
   cy.intercept("POST", "/api/dataset").as("dataset");
 
   cy.createNativeQuestion(
     {
-      name: "16386",
+      name: questionName,
       native: {
         query: BOOLEAN_QUERY,
       },
@@ -21,4 +18,7 @@ export const setupBooleanQuery = () => {
 
   cy.findByText("Explore results").click();
   cy.wait("@dataset");
-};
+}
+
+const BOOLEAN_QUERY =
+  'select 0::integer as "integer", true::boolean AS "boolean" union all \nselect 1::integer as "integer", false::boolean AS "boolean" union all \nselect null as "integer", true::boolean AS "boolean" union all \nselect -1::integer as "integer", null AS "boolean"';
